\name{get_iseq}
\alias{get_iseq}
\docType{package}
\title{
Get Individual Disease Progression Sequence
}
\description{

}
\usage{
get_iseq_r(x, df, 
           var_id = "ID", var_t = "date", 
           var_diag = c("ICD_1","ICD_2","ICD_3","ICD_4","ICD_5"))	
}
\arguments{
  \item{x}{An id}

  \item{df}{A dataframe}
}
\value{

}
\examples{
\dontrun{

}
}
