\name{iseq_els}
\alias{iseq_els}
\docType{package}
\title{
Get progression links from Progression Sequence
}
\description{
Simple function using Rcpp
}
\usage{
iseq_els(t, dz, drop_bilink = FALSE)
}

\arguments{
  \item{t}{onset time}
  \item{dz}{type of disease}
  \item{drop_bilink}{whether to ignore bidirectional links}
}

\examples{
\dontrun{

}
}
